HV: * Upped the compiler warninglevels a bit more, this found       a couple subtle issues as well a number slightly more       serious. implicit conversions are mostly mode explicit now.       Added a few const_casts<>. Fixed some types (unsigned int -> int)       because the systemcalls expect int. Fixed a number of template <>       specialization errors. double comparisions using '==' changed       into 'near()'.     * default optimization level is now -O2     * FIXED off-by-one bug in "block.sub()" method, you couldn't       extract the whole block or up-to-and-including the last byte,       more specifically.     * Added support for Mark5C detection and flags.     * Added another signature in thunk.h [member function of const       object, taking 1 argument]     * Added support for Bankmode switching to xlrdev.

+++ evlbi5a/bittwiddlinghacks.h
@@ -55,7 +55,7 @@ inline unsigned int count_bits(uint64_t v) {
     v += v >>  8;
     v += v >> 16;
     v += v >> 32;
-    return v & 0x7f;
+    return (unsigned int)(v & 0x7f);
 }
 
 // Count the number of 'edges' or transitions from 0->1 and 1->0 when
